% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgScoreByCohort.R
\name{avgScoreByCohort}
\alias{avgScoreByCohort}
\title{Average Score by Cohort}
\usage{
avgScoreByCohort(df)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes a line level dataset of patient data and
calculates the average score for each cohort
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}
}
